Updated cell detection and avg of vols/amps
Changed the levels for cell-detection to avoid false detection of to
many cells. The wrong number of cells may be detected if the battery is
low on powerup. Won't change the cell count in flight if voltage drops.
Changed the averaging of volts/amps to report the average volts/amps
received through mavlink between each poll.
Added a delay to try to eliminate low battery-warnings on model powerup.